SETTING: add setting for Postgresql. Fix: fix models on Drones App. Add: add migrations for Drones app.
