+++ requirements.txt
@@ -39,8 +39,6 @@ Pygments==2.5.2
 pyrsistent==0.15.7
 python-dateutil==2.8.1
 pytz==2019.3
-pywin32==227
-pywinpty==0.5.7
 pyzmq==18.1.1
 qtconsole==4.6.0
 Send2Trash==1.5.0
